scripts/requirements.txt
pypdfium2>=4.0
google-generativeai>=0.8
supabase>=2.0
python-dotenv>=1.0
numpy>=1.24
